# contactapi/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "STIS Contact API",
    "description": "API ini menyediakan layanan untuk mengelola dan mengakses informasi kontak yang dimiliki oleh mahasiswa dan dosen di Politeknik Statistika STIS. Layanan ini dirancang untuk mempermudah pencarian informasi kontak dengan menyediakan berbagai endpoint untuk menambah, memperbarui, dan menghapus kontak dari mahasiswa dan dosen. API ini juga mendukung pengelompokan kontak berdasarkan organisasi mahasiswa atau mata kuliah yang diampu oleh dosen serta menawarkan fitur pencarian yang fleksibel. \n\nFitur utama meliputi: \n- Manajemen kontak mahasiswa dan dosen \n- Pengelompokan kontak berdasarkan organisasi atau jurusan \n- Dukungan endpoint untuk CRUD (Create, Read, Update, Delete) pada data kontak \n- Fitur pencarian kontak berdasarkan organisasi atau mata kuliah yang diampu oleh dosen\n\n",
    "contact": {
      "name": "Dodi Firmansyah",
      "url": "https://github.com/dodfrm",
      "email": "222212572@stis.ac.id"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "operationId": "getContactById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Contact found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update an existing contact",
        "operationId": "updateContact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "fullName": "Dodi Firmansyah",
                  "phone": "081998915680",
                  "email": "222212572@stis.ac.id",
                  "contactType": "MAHASISWA or DOSEN"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete a contact",
        "operationId": "deleteContact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Contact deleted successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "responses": {
          "200": {
            "description": "List of all contacts",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "fullName": "Dodi Firmansyah",
                  "phone": "081998915680",
                  "email": "222212572@stis.ac.id",
                  "contactType": "MAHASISWA or DOSEN"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contacts/search": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Search contact by criteria",
        "operationId": "searchContacts",
        "parameters": [
          {
            "name": "fullName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jabatan",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["KETUA", "WAKIL_KETUA", "BENDAHARA", "PJ_MATKUL"]
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["MAHASISWA", "DOSEN"]
            }
          },
          {
            "name": "organizationName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subjectName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Contact not found"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Contact found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subjectName": {
            "type": "string"
          }
        }
      },
      "ContactSubject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          }
        }
      },
      "ContactOrganization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kelas": {
            "type": "string"
          },
          "jabatan": {
            "type": "string",
            "enum": ["KETUA", "WAKIL_KETUA", "BENDAHARA", "PJ_MATKUL"]
          },
          "periodeJabatan": {
            "type": "string"
          }
        }
      },
...
      "ContactDto": {
        "required": ["fullName", "phone"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contactType": {
            "type": "string",
            "enum": ["MAHASISWA", "DOSEN"]
          },
          "contactOrganizations": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactOrganization"
            }
          },
          "contactSubjects": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactSubject"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Input token sebagai 'Bearer \u003Ctoken\u003E'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}